Implements Vault Helpers
