+++ src/charm.py
@@ -13,7 +13,6 @@ import socket
 from datetime import datetime
 from typing import Any, Dict, List
 
-import hcl
 from botocore.response import StreamingBody
 from charms.data_platform_libs.v0.s3 import S3Requirer
 from charms.grafana_k8s.v0.grafana_dashboard import GrafanaDashboardProvider
@@ -38,12 +37,17 @@ from charms.vault_k8s.v0.vault_autounseal import (
     VaultAutounsealRequires,
 )
 from charms.vault_k8s.v0.vault_client import (
-    AppRole,
     AuditDeviceType,
     SecretsBackend,
     Token,
     VaultClient,
     VaultClientError,
+)
+from charms.vault_k8s.v0.vault_helpers import (
+    VaultHelpers,
+    config_file_content_matches,
+    render_vault_config_file,
+    seal_type_has_changed,
 )
 from charms.vault_k8s.v0.vault_kv import (
     VaultKvClientDetachedEvent,
@@ -60,7 +64,6 @@ from charms.vault_k8s.v0.vault_managers import (
     VaultCertsError,
 )
 from charms.vault_k8s.v0.vault_s3 import S3, S3Error
-from jinja2 import Environment, FileSystemLoader
 from ops import CharmBase, MaintenanceStatus, main
 from ops.charm import (
     ActionEvent,
@@ -90,7 +93,6 @@ BACKUP_KEY_PREFIX = "vault-backup"
 CHARM_POLICY_NAME = "charm-access"
 CHARM_POLICY_PATH = "src/templates/charm_policy.hcl"
 CONFIG_TEMPLATE_DIR_PATH = "src/templates/"
-CONFIG_TEMPLATE_NAME = "vault.hcl.j2"
 CONTAINER_NAME = "vault"
 CONTAINER_TLS_FILE_DIRECTORY_PATH = "/vault/certs"
 KV_RELATION_NAME = "vault-kv"
@@ -104,7 +106,6 @@ PROMETHEUS_ALERT_RULES_PATH = "./src/prometheus_alert_rules"
 REQUIRED_S3_PARAMETERS = ["bucket", "access-key", "secret-key", "endpoint"]
 S3_RELATION_NAME = "s3-parameters"
 TLS_CERTIFICATES_PKI_RELATION_NAME = "tls-certificates-pki"
-VAULT_CHARM_APPROLE_SECRET_LABEL = "vault-approle-auth-details"
 VAULT_CONFIG_FILE_PATH = "/vault/config/vault.hcl"
 VAULT_STORAGE_PATH = "/vault/raft"
 
@@ -123,6 +124,7 @@ class VaultCharm(CharmBase):
     def __init__(self, *args: Any):
         super().__init__(*args)
         self.juju_facade = JujuFacade(self)
+        self.helpers = VaultHelpers(self)
         self._service_name = self._container_name = CONTAINER_NAME
         self._container = Container(container=self.unit.get_container(self._container_name))
         self.unit.set_ports(self.VAULT_PORT)
@@ -223,7 +225,7 @@ class VaultCharm(CharmBase):
         """Handle the collect status event."""
         if (
             self.juju_facade.relation_exists(TLS_CERTIFICATES_PKI_RELATION_NAME)
-            and not self._common_name_config_is_valid()
+            and not self.helpers.common_name_config_is_valid()
         ):
             event.add_status(
                 BlockedStatus(
@@ -278,12 +280,17 @@ class VaultCharm(CharmBase):
         except VaultClientError:
             event.add_status(MaintenanceStatus("Seal check failed, waiting for Vault to recover"))
             return
-        if not self._get_approle_auth_secret():
+        if not self.helpers.get_approle_auth_secret():
             event.add_status(
                 BlockedStatus("Please authorize charm (see `authorize-charm` action)")
             )
             return
-        if not self._get_active_vault_client():
+        try:
+            vault = self.helpers.get_active_vault_client(
+                self._api_address, self.tls.get_tls_file_path_in_charm(File.CA)
+            )
+        except VaultCertsError as e:
+            logger.warning("Waiting for Vault Client %s", e)
             event.add_status(WaitingStatus("Waiting for vault to finish raft leader election"))
         event.add_status(ActiveStatus())
 
@@ -311,22 +318,15 @@ class VaultCharm(CharmBase):
         self._generate_vault_config_file()
         self._set_pebble_plan()
         try:
-            vault = VaultClient(
-                url=self._api_address, ca_cert_path=self.tls.get_tls_file_path_in_charm(File.CA)
-            )
+            if not (
+                vault := self.helpers.get_active_and_unsealed_vault_client(
+                    self._api_address, self.tls.get_tls_file_path_in_charm(File.CA)
+                )
+            ):
+                logger.warning("Failed to get Vault Client.")
+                return
         except VaultCertsError as e:
             logger.error("Failed to get TLS file path: %s", e)
-            return
-        if not vault.is_api_available():
-            return
-        if not vault.is_initialized():
-            return
-        try:
-            if vault.is_sealed():
-                return
-        except VaultClientError:
-            return
-        if not (vault := self._get_active_vault_client()):
             return
         self._configure_pki_secrets_engine(vault)
         self._sync_vault_autounseal(vault)
@@ -358,19 +358,17 @@ class VaultCharm(CharmBase):
 
     def _remove_node_from_raft_cluster(self):
         """Remove the node from the raft cluster."""
-        if not (approle := self._get_approle_auth_secret()):
-            return
-        vault = VaultClient(url=self._api_address, ca_cert_path=None)
-        if not vault.is_api_available():
-            return
-        if not vault.is_initialized():
-            return
         try:
-            if vault.is_sealed():
+            if not (
+                vault := self.helpers.get_active_and_unsealed_vault_client(
+                    self._api_address, self.tls.get_tls_file_path_in_charm(File.CA)
+                )
+            ):
+                logger.warning("Failed to get Vault Client.")
                 return
-        except VaultClientError:
+        except VaultCertsError as e:
+            logger.error("Failed to get TLS file path: %s", e)
             return
-        vault.authenticate(approle)
         if vault.is_node_in_raft_peers(self._node_id) and vault.get_num_raft_peers() > 1:
             vault.remove_raft_node(self._node_id)
 
@@ -399,6 +397,7 @@ class VaultCharm(CharmBase):
             is_ca=True,
         )
 
+    # TODO It can be refactored to the feature manager
     def _sync_vault_autounseal(self, vault_client: VaultClient) -> None:
         """Sync the vault autounseal relation."""
         if not self.unit.is_leader():
@@ -510,11 +509,7 @@ class VaultCharm(CharmBase):
                 token_max_ttl="1h",
             )
             secret_id = vault.generate_role_secret_id(name=APPROLE_ROLE_NAME, cidrs=cidrs)
-            self.juju_facade.set_app_secret_content(
-                content={"role-id": role_id, "secret-id": secret_id},
-                label=VAULT_CHARM_APPROLE_SECRET_LABEL,
-                description="The authentication details for the charm's access to vault.",
-            )
+            self.helpers.set_approle_auth_secret(role_id, secret_id)
             event.set_results(
                 {"result": "Charm authorized successfully. You may now remove the secret."}
             )
@@ -555,7 +550,14 @@ class VaultCharm(CharmBase):
             logger.error("Failed to run create-backup action - Failed to create S3 bucket.")
             return
         backup_key = self._get_backup_key()
-        vault = self._get_active_vault_client()
+        try:
+            vault = self.helpers.get_active_vault_client(
+                self._api_address, self.tls.get_tls_file_path_in_charm(File.CA)
+            )
+        except VaultCertsError as e:
+            event.fail(message="Failed to initialize Vault client.")
+            logger.error("Failed to run create-backup action - %s", e)
+            return
         if not vault:
             event.fail(message="Failed to initialize Vault client.")
             logger.error("Failed to run create-backup action - Failed to initialize Vault client.")
@@ -662,17 +664,15 @@ class VaultCharm(CharmBase):
             event.fail(message="Failed to restore vault.")
             return
         try:
-            if self.juju_facade.secret_exists_with_fields(
-                fields=("role-id", "secret-id"),
-                label=VAULT_CHARM_APPROLE_SECRET_LABEL,
-            ):
-                approle = self._get_approle_auth_secret()
+            if self.helpers.get_approle_auth_secret():
+                approle = self.helpers.get_approle_auth_secret()
+                # TODO, don't we need an active vault here?
                 vault = VaultClient(
                     url=self._api_address,
                     ca_cert_path=self.tls.get_tls_file_path_in_charm(File.CA),
                 )
                 if approle and not vault.authenticate(approle):
-                    self.juju_facade.remove_secret(label=VAULT_CHARM_APPROLE_SECRET_LABEL)
+                    self.helpers.remove_approle_auth_secret()
         except Exception as e:
             logger.error("Failed to remove old approle secret: %s", e)
             event.fail(message="Failed to remove old approle secret.")
@@ -741,11 +741,6 @@ class VaultCharm(CharmBase):
             return None
         return f"https://{ingress_address}:{self.VAULT_PORT}"
 
-    def _common_name_config_is_valid(self) -> bool:
-        """Return whether the config value for the common name is valid."""
-        common_name = self.juju_facade.get_string_config("common_name")
-        return common_name != ""
-
     def _generate_vault_config_file(self) -> None:
         """Handle the creation of the Vault config file."""
         retry_joins = [
@@ -758,7 +753,7 @@ class VaultCharm(CharmBase):
 
         autounseal_configuration_details = self._get_vault_autounseal_configuration()
 
-        content = _render_vault_config_file(
+        content = render_vault_config_file(
             default_lease_ttl=self.juju_facade.get_string_config("default_lease_ttl"),
             max_lease_ttl=self.juju_facade.get_string_config("max_lease_ttl"),
             cluster_address=self._cluster_address,
@@ -781,7 +776,7 @@ class VaultCharm(CharmBase):
             # If the seal type has changed, we need to restart Vault to apply
             # the changes. SIGHUP is currently only supported as a beta feature
             # for the enterprise version in Vault 1.16+
-            if _seal_type_has_changed(existing_content, content):
+            if seal_type_has_changed(existing_content, content):
                 if self._vault_service_is_running():
                     self._container.restart(self._service_name)
 
@@ -808,23 +803,6 @@ class VaultCharm(CharmBase):
         """Push the config file to the workload."""
         self._container.push(path=VAULT_CONFIG_FILE_PATH, source=content)
         logger.info("Pushed %s config file", VAULT_CONFIG_FILE_PATH)
-
-    def _get_approle_auth_secret(self) -> AppRole | None:
-        """Get the vault approle login details secret.
-
-        Returns:
-            AppRole: An AppRole object with role_id and secret_id set from the
-                     values stored in the Juju secret, or None if the secret is
-                     not found or either of the values are not set.
-        """
-        try:
-            role_id, secret_id = self.juju_facade.get_secret_content_values(
-                "role-id", "secret-id", label=VAULT_CHARM_APPROLE_SECRET_LABEL
-            )
-        except NoSuchSecretError:
-            logger.warning("Approle secret not yet created")
-            return None
-        return AppRole(role_id, secret_id) if role_id and secret_id else None
 
     def _get_missing_s3_parameters(self) -> List[str]:
         """Return the list of missing S3 parameters.
@@ -861,7 +839,7 @@ class VaultCharm(CharmBase):
             logger.error("Failed to find active Vault client, cannot restore snapshot.")
             return False
         try:
-            if not (approle := self._get_approle_auth_secret()):
+            if not (approle := self.helpers.get_approle_auth_secret()):
                 logger.error("Failed to log in to Vault")
                 return False
             vault.authenticate(approle)
@@ -878,33 +856,12 @@ class VaultCharm(CharmBase):
 
         return True
 
-    def _get_active_vault_client(self) -> VaultClient | None:
-        """Return an initialized vault client.
-
-        Returns:
-            Vault: An active Vault client configured with the cluster address
-                   and CA certificate, and authorized with the AppRole
-                   credentials set upon initial authorization of the charm, or
-                   `None` if the client could not be successfully created or
-                   has not been authorized.
-        """
-        try:
-            vault = VaultClient(
-                url=self._api_address,
-                ca_cert_path=self.tls.get_tls_file_path_in_charm(File.CA),
-            )
-        except VaultCertsError as e:
-            logger.warning("Failed to get Vault client: %s", e)
-            return None
-        if not vault.is_api_available():
-            return None
-        if not (approle := self._get_approle_auth_secret()):
-            return None
-        if not vault.authenticate(approle):
-            return None
-        if not vault.is_active_or_standby():
-            return None
-        return vault
+    def _get_peer_node_api_addresses(self) -> List[str]:
+        """Return a list of unit addresses that should be a part of the raft cluster."""
+        return [
+            f"https://{self.app.name}-{i}.{socket.getfqdn().split('.', 1)[-1]}:{self.VAULT_PORT}"
+            for i in range(self.app.planned_units())
+        ]
 
     @property
     def _vault_layer(self) -> Layer:
@@ -923,13 +880,6 @@ class VaultCharm(CharmBase):
                 },
             }
         )
-
-    def _get_peer_node_api_addresses(self) -> List[str]:
-        """Return a list of unit addresses that should be a part of the raft cluster."""
-        return [
-            f"https://{self.app.name}-{i}.{socket.getfqdn().split('.', 1)[-1]}:{self.VAULT_PORT}"
-            for i in range(self.app.planned_units())
-        ]
 
     @property
     def _node_id(self) -> str:
@@ -972,95 +922,6 @@ class VaultCharm(CharmBase):
             str: Ingress address
         """
         return self.juju_facade.get_ingress_address(PEER_RELATION_NAME)
-
-
-def _render_vault_config_file(
-    default_lease_ttl: str,
-    max_lease_ttl: str,
-    cluster_address: str,
-    api_address: str,
-    tls_cert_file: str,
-    tls_key_file: str,
-    tcp_address: str,
-    raft_storage_path: str,
-    node_id: str,
-    retry_joins: List[Dict[str, str]],
-    autounseal_details: AutounsealConfigurationDetails | None = None,
-) -> str:
-    jinja2_environment = Environment(loader=FileSystemLoader(CONFIG_TEMPLATE_DIR_PATH))
-    template = jinja2_environment.get_template(CONFIG_TEMPLATE_NAME)
-    content = template.render(
-        default_lease_ttl=default_lease_ttl,
-        max_lease_ttl=max_lease_ttl,
-        cluster_address=cluster_address,
-        api_address=api_address,
-        tls_cert_file=tls_cert_file,
-        tls_key_file=tls_key_file,
-        tcp_address=tcp_address,
-        raft_storage_path=raft_storage_path,
-        node_id=node_id,
-        retry_joins=retry_joins,
-        autounseal_address=autounseal_details.address if autounseal_details else None,
-        autounseal_key_name=autounseal_details.key_name if autounseal_details else None,
-        autounseal_mount_path=autounseal_details.mount_path if autounseal_details else None,
-        autounseal_token=autounseal_details.token if autounseal_details else None,
-        autounseal_tls_ca_cert=autounseal_details.ca_cert_path if autounseal_details else None,
-    )
-    return content
-
-
-def _seal_type_has_changed(content_a: str, content_b: str) -> bool:
-    """Check if the seal type has changed between two versions of the Vault configuration file.
-
-    Currently only checks if the transit stanza is present or not, since this
-    is all we support. This function will need to be extended to support
-    alternate cases if and when we support them.
-    """
-    config_a = hcl.loads(content_a)
-    config_b = hcl.loads(content_b)
-    return _contains_transit_stanza(config_a) != _contains_transit_stanza(config_b)
-
-
-def _contains_transit_stanza(config: dict) -> bool:
-    return "seal" in config and "transit" in config["seal"]
-
-
-def config_file_content_matches(existing_content: str, new_content: str) -> bool:
-    """Return whether two Vault config file contents match.
-
-    We check if the retry_join addresses match, and then we check if the rest of the config
-    file matches.
-
-    Returns:
-        bool: Whether the vault config file content matches
-    """
-    existing_config_hcl = hcl.loads(existing_content)
-    new_content_hcl = hcl.loads(new_content)
-    if not existing_config_hcl:
-        logger.info("Existing config file is empty")
-        return existing_config_hcl == new_content_hcl
-    if not new_content_hcl:
-        logger.info("New config file is empty")
-        return existing_config_hcl == new_content_hcl
-
-    new_retry_joins = new_content_hcl["storage"]["raft"].pop("retry_join", [])
-    existing_retry_joins = existing_config_hcl["storage"]["raft"].pop("retry_join", [])
-
-    # If there is only one retry join, it is a dict
-    if isinstance(new_retry_joins, dict):
-        new_retry_joins = [new_retry_joins]
-    if isinstance(existing_retry_joins, dict):
-        existing_retry_joins = [existing_retry_joins]
-
-    new_retry_join_api_addresses = {address["leader_api_addr"] for address in new_retry_joins}
-    existing_retry_join_api_addresses = {
-        address["leader_api_addr"] for address in existing_retry_joins
-    }
-
-    return (
-        new_retry_join_api_addresses == existing_retry_join_api_addresses
-        and new_content_hcl == existing_config_hcl
-    )
 
 
 if __name__ == "__main__":  # pragma: no cover
